update rest api structure
