+++ Utils/hdinsight-node-common/src/com/microsoft/azure/hdinsight/sdk/rest/spark/Application.java
@@ -19,19 +19,57 @@
  * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.microsoft.azure.hdinsight.sdk.jobs.yarn;
+package com.microsoft.azure.hdinsight.sdk.rest.spark;
 
-/**
- * Created by ltian on 5/6/2017.
- */
-public class YarnApplications {
-    private YarnApplication apps;
 
-    public YarnApplication getApps() {
-        return apps;
+import com.microsoft.azure.hdinsight.sdk.rest.AttemptWithAppId;
+import com.microsoft.azure.hdinsight.sdk.rest.IConvertible;
+
+import java.util.List;
+
+public class Application implements IConvertible
+{
+    private String id;
+
+    private List<Attempt> attempts;
+
+    private String name;
+
+    public String getId ()
+    {
+        return id;
     }
 
-    public void setApps(YarnApplication apps) {
-        this.apps = apps;
+    public void setId (String id)
+    {
+        this.id = id;
+    }
+
+    public List<Attempt> getAttempts ()
+    {
+        return attempts;
+    }
+
+    public void setAttempts (List<Attempt> attempts)
+    {
+        this.attempts = attempts;
+    }
+
+    public String getName ()
+    {
+        return name;
+    }
+
+    public void setName (String name)
+    {
+        this.name = name;
+    }
+
+    public AttemptWithAppId getLastAttemptWithAppId() {
+        final int attemptTimes = attempts.size();
+        Attempt lastAttempt = attempts.stream().filter(attempt -> {
+            return Integer.valueOf(attempt.getAttemptId()) == attemptTimes;
+        }).findFirst().get();
+        return new AttemptWithAppId("spark2withblob",getId(), lastAttempt);
     }
 }
