+++ Utils/hdinsight-node-common/src/com/microsoft/azure/hdinsight/sdk/rest/spark/Attempt.java
@@ -19,13 +19,11 @@
  * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.microsoft.azure.hdinsight.sdk.jobs.spark;
+package com.microsoft.azure.hdinsight.sdk.rest.spark;
 
+import com.microsoft.azure.hdinsight.sdk.rest.IConvertible;
 
-/**
- * Created by ltian on 5/6/2017.
- */
-public class Attempt {
+public class Attempt implements IConvertible {
     private String startTime;
 
     private String sparkUser;
